% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2l.R
\name{d2l}
\alias{d2l}
\title{Distance to Line}
\usage{
d2l(x, y, k = 1, h = 0)
}
\arguments{
\item{x}{(numeric) X-coordinates.}

\item{y}{(numeric) Y-coordinates.}

\item{k}{(numeric) The k for the target line y=kx+h.}

\item{h}{(numeric) The h for the target line y=kx+h.}
}
\value{
(numeric) Distance values.
}
\description{
In 2D space, calculate the distance from points to a target line.
}
\examples{

x <- c(1, 2)
y <- c(1, 3)
d2l(x, y, 1, 0)

}
\author{
DING, HONGXU (dinghongxu1130@gmail.com)
}
