% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{Fisher's P-Value Integration}
\usage{
fisher(p)
}
\arguments{
\item{p}{(numeric) P-vales.}
}
\value{
(numeric) The integrated p-value using Fisher's method.
}
\description{
Perform Fisher's method for p-value integration.
}
\examples{

p <- c(0.001, 0.01, 0.1)
fisher(p)

}
\author{
DING, HONGXU (dinghongxu1130@gmail.com)
}
