% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2l.R
\name{p2l}
\alias{p2l}
\title{Projection to Line}
\usage{
p2l(x, y, k, h)
}
\arguments{
\item{x}{(numeric) X-coordinates.}

\item{y}{(numeric) Y-coordinates.}

\item{k}{(numeric) The k for the target y=kx+h.}

\item{h}{(numeric) The h for the target y=kx+h.}
}
\value{
(numeric) Projection coordinates on the target line.
}
\description{
In 2D space, calculate the projection coordinates of points to a line.
}
\examples{

x <- c(1, 2)
y <- c(2, 3)
k <- 1
h <- 0
pj <- p2l(x, y, k, h)
plot(x, y, xlim = c(0, 5), ylim = c(0, 5))
abline(h, k, NULL, NULL)
arrows(x[1], y[1], pj[1, 1], pj[1, 2])
arrows(x[2], y[2], pj[2, 1], pj[2, 2])

}
\author{
DING, HONGXU (dinghongxu1130@gmail.com)
}
