% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum}
\alias{spectrum}
\title{Color Spectrum}
\usage{
spectrum(
  x,
  pcol = c("grey", "red"),
  ncol = c("grey", "green"),
  range = 5,
  resolution = 100
)
}
\arguments{
\item{x}{(numeric) The numeric vector to be represented as a color spectrum.}

\item{pcol}{(character) Colors for positive values.}

\item{ncol}{(character) Colors for negative values.}

\item{range}{(numeric) Xs with freater absolute value will be set as the "maximum" pcol/ncol.}

\item{resolution}{(numeric) The number of interpolations for pcol/ncol.}
}
\value{
(character) The color spectrum vector.
}
\description{
Create a spectrum of colors corresponded to a numeric vector.
}
\examples{

x <- 1:21
y <- 1:21
color <- spectrum(-10:10)
plot(x, y, col = color, pch = 16)

}
\author{
DING, HONGXU (dinghongxu1130@gmail.com)
}
