% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssde.R
\name{ssde}
\alias{ssde}
\title{Single Sample Differential Expression}
\usage{
ssde(x, y, k = NULL, h = NULL, neighbor = 200, epoch = 1000, seed = 1)
}
\arguments{
\item{x}{(numeric) X-coordinates.}

\item{y}{(numeric) Y-coordinates.}

\item{k}{(numeric) The k for the target line y=kx+h.}

\item{h}{(numeric) The h for the target line y=kx+h.}

\item{neighbor}{(numeric) The number of neighboring genes for calculating empirical p-values.}

\item{epoch}{(numeric) The number of permutations for building null models for calculating empirical p-values.}
}
\value{
(list) A list containing the k, h, d and p values, as the ssde result.
}
\description{
The main function for performing single sample differential expression analysis.
}
\examples{

library(bcellViper)
data(bcellViper)
x <- dset@assayData$exprs[, "GSM44113"]
y <- dset@assayData$exprs[, "GSM44246"]
res <- ssde(x, y, k=1, h=0, epoch=100)

}
\author{
DING, HONGXU (dinghongxu1130@gmail.com)
}
